Update README to be more general
The previous version of the README was specific to the University of the South Pacific (USP). The update makes the wording more generalized, making it applicable to students of any university or institution. The repository is now more inclusive and no longer specifically targeted towards USP students, allowing more diverse contributions.